Added scroll functionality to the CardList; Organized files; Refactored the code in App.js.
