--- src/App.js
@@ -1,45 +1,0 @@
-import React, {Component} from 'react';
-import CardList from './CardList';
-import SearchBox from './SearchBox';
-
-class App extends Component {
-    constructor() {
-        super();
-        this.state = {
-            robots: [],
-            searchfield: ''
-        }
-    }
-
-    componentDidMount() {
-        fetch('https://jsonplaceholder.typicode.com/users')
-            .then(Response => Response.json())
-            .then(users => {this.setState({robots: users})});
-    }
-
-    eSearchChange = (event) => {
-        this.setState({searchfield: event.target.value});
-    }
-
-    render() {
-
-        const filter = this.state.robots.filter(robot => {
-            return robot.name.toLowerCase().includes(this.state.searchfield.toLowerCase());
-        });
-        
-        if (this.state.robots.length === 0) {
-            return <h1>Loading</h1>
-        }
-        else {
-            return (
-                <div className='tc'>
-                    <h1 className='f1'>RoboFriends</h1>
-                    <SearchBox searchChange={this.eSearchChange} />
-                    <CardList robots={filter}/>
-                </div>
-            );
-        }
-    }
-}
-
-export default App;
